eval one attribute at a time
previously, we were simultaneously eval-ing all cells.
this led to all sorts of weird race condition behavior,
because when we eval a cell its deps might not be done yet.
now, instead we eval one attr at a time, but we eval
all rows concurrently. this has a few nice properties
and solves a bunch of problems:

- rows don't depend on each other, so it's safe
- if a column does network i/o, it happens concurrently
- we only trigger a small number of table update events
  (one per column) which avoids performance issues related to
  overloading the redux event count

there are two pending issues to resolve still:

- we re-eval everything every time anything changes;
  we need some memoization (maybe do this at the level of
  individual functions rather than the runtime?)
- we eval things in a silly order that doesn't capture deps,
  so updates to columns don't flow correctly right-to-left